Working on the Nav bar and its functionality
