+++ components/layout/layout.component.js
@@ -11,7 +11,11 @@ export const Layout = ({ children, name, desc }) => {
         <title>Mark-Bond Konsult - {name}</title>
         <link rel="icon" href="/logo/errands_fav.jpeg" />
         <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossOrigin />
+        <link
+          rel="preconnect"
+          href="https://fonts.gstatic.com"
+          crossOrigin="true"
+        />
         <link
           href="https://fonts.googleapis.com/css2?family=Roboto+Slab:wght@300;400;500;600;700;800;900&family=Roboto:wght@100;300;400;500;700;900&display=swap"
           rel="stylesheet"
